Rename anon folder to aggregation
