+++ pg_diffix/aggregation/random.h
