+++ src/aggregation/contribution_tracker.c
@@ -3,7 +3,7 @@
 #include "utils/elog.h"
 
 #include "pg_diffix/config.h"
-#include "pg_diffix/anon/contribution_tracker.h"
+#include "pg_diffix/aggregation/contribution_tracker.h"
 
 /* ----------------------------------------------------------------
  * Top contributors management
