+++ src/aggregation/lcf.c
@@ -6,8 +6,8 @@
 #include <inttypes.h>
 
 #include "pg_diffix/config.h"
-#include "pg_diffix/anon/aid_tracker.h"
-#include "pg_diffix/anon/random.h"
+#include "pg_diffix/aggregation/aid_tracker.h"
+#include "pg_diffix/aggregation/random.h"
 
 typedef struct LcfResult
 {
