+++ src/aggregation/random.c
@@ -4,7 +4,7 @@
 #include <math.h>
 
 #include "pg_diffix/config.h"
-#include "pg_diffix/anon/random.h"
+#include "pg_diffix/aggregation/random.h"
 
 extern double pg_erand48(unsigned short xseed[3]);
